refactor(organize): move organize closures to organize_cli module
Six inner closures (confirm, move-tracking, ASK-mode date reads, ASK
plan-item execution, ASK execution, interactive conflict handling)
become module-level functions in fx_bin/organize_cli.py with explicit
parameters. cli.py drops 207 lines.

One signature beyond the captured-variable list: tests patch
fx_bin.cli.sys (module-name swap) to fake a TTY, which cannot reach an
isatty() call that moved modules — cli.py now computes
sys.stdin.isatty() at each call site and passes is_tty explicitly.
Behavior identical (same check, same logical point); zero test edits.

The bare `raise` in the ASK execution loop was determined to be
unreachable (err_delta==1 is only ever returned when fail_fast is
False) and is moved verbatim; removal belongs to the dead-branch task.
